Add an environment variable to pass through an access token
Check the `GOOGLE_ACCESS_TOKEN` environment variable to for a static
token before falling back to the normal routine. This could allow a user
to pass through a one-time token when building with limited scopes and
which expires soon. This provides increased security in situations where
it is possible to use such a token over full credentials.